# Combine some commands to make docker launches faster.