Adds a PTY to show Docker push etc status and merges some commands to speed launches.

Before:

touch config/0; python infra/launch.py --foreground --tpu_name=test-spin-up-8  --  --version  7.16s user 1.22s system 31% cpu 26.867 total

After:

touch config/1; python infra/launch.py --foreground --tpu_name=test-spin-up-8  --  --version  3.65s user 0.67s system 29% cpu 14.765 total